Added method to calculate beam vectors from array of miller indices and rotation angles.
Added method to calculated detector coordinates from array of beam vectors.

Since this can fail and throw an exception, the latter needs to be modified to show which transforms failed. At the moment is sets the detector coordinate to (-1, -1). Possibly have a look at how numpy masked arrays are implemented or see how to pass flex arrays by reference (tried this but not straight forward)