Tidied up some stuff: put array specific functions into a dials/array_family folder.
Created a class to calculate the reflection mask roi for each reflection based on the standard deviation of the beam divergence and mosaicity in the xds coordinate system.
Had a bit of a problem. It seems scitbx::af::tiny <int, 6> does not map transparently onto a tuple but scitbx::af::tiny <double, 6> does. I'm probably doing something wrong but I've made a note to have a look later; in any case the function currently returns the roi as double rather than int pixel coordinates.