Modified goniometer method for getting frame number from angle to have option to either wrap around or not.
The two cases for this are, when generating the rotation angles we want to get the valid frame number for that rotation.

In calculating the reflection mask, we want to calculate the range of frames over which the reflection is recorded (which could extend to negative frames) in which case wrapping the angle gives funny results.