Made some changes to inheritance test.
In the previous test, BaseWrapper was imported into python as "Base".

This meant that when I just wanted as Base object and passed it to a function calling it's methods, I would have the added overhead of the boost python stuff.

I don't this this is a good solution, using the c++ objects directly shouldn't incur extra overhead.

Therefore, I've split the test objects into export Base and Derived directly and have exposed the BaseWrapper (and DerivedWrapper) into python as themselves (changed names to BaseInterface, DerivedInterface). 

Python classes would have to inherit from these instead of directly from Base, Derived. It's not as neat but means you only pay for what you get in terms of extra boost python overhead.