Added centroid_position and centroid_variance to reflection class. Fixed bug in reflection_profile_helpers. Modified centroid interface to take reflection list. Modified toy_centroid_runner to use new centroid interface.
Note that no filtering is done to the reflections at the moment to ensure that we're only using reflections within the image volume or that the refelction only covers a valid area of the detector. This means that some reflection images will contain pixels with -1 counts etc.

The changes that need to be made to the centroid implementations are shown in toy_centroid which now uses the new interface.

To run toy_centroid_runner use a command like the following:

dials.toy_centroid_runner ./GXPARM.XDS ./INTEGRATE.HKL ./X1_strong_M1S1_1_0{001..300}.cbf