Do not scale the objective to [0,1].
This was only done for the convergence test to compare against a fixed gradient threshold. Instead we can scale that threshold to the appropriate value. The minimisation is unaffected (apart from differences due to floating point precision), but there are advantages:
* The problem retains its natural scale. It is easier to compare results.
* The weights remain on the same scale as the residuals.
* It is easier to calculate things like the errors on parameter estimates as we don't need to remember an additional scale factor (improved clarity).

NB The use of an arbitrary gradient_threshold is diasgreeable and a better convergence criterion should be found. The current mode of adapting the gradient_threshold based on the initial value of the objective means we are chasing a moving target: a second round of minimisation will adapt to a tighter threshold. A better criterion would use the rate of reduction in rmsd.