Slightly hacky test of multi panel detector parameterisation.
A refinement run on synthetic data is performed twice, first with
a single panel detector parameterisation, and secondly with a
3x3 array of panels that as a group fill exactly the space of the
single panel detector. Between runs the parameter values are
rewound so that the runs should be identical.

The test finishes by checking that aspects of the two runs really
were identical, including rmsd values per step and parameter
vector values per step.