More informative test for scan-varying reflection predictor.
Reflections are predicted both by the usual predictor and by the
scan-varying predictor, for a scan during which UB does not
change. We would expect the same reflections to be predicted, with
slightly different reflecting angle due to the approximation
made by the scan-varying predictor.

Indeed, the same number of reflections are predicted in each case.
The reflection lists are sorted, then individual reflections are
compared. They are the same reflections, with the rotation angle
from scan-varying prediction founs to be within 1% of the image
width (1 degree) from the more accurate value determined by the
normal predictor.