go_fast for scan-varying refinement. The gains are not quite as dramatic as for the scan-static case, but there is still scope for improvement.
Note, the first round of prediction and gradient calculation is
done before any outlier rejection or random sampling. This can
take a while as all reflections are considered. It is probably
better if some sampling is done before scan-varying refinement.
One possibility would be to use as input only the reflections that
were used during a preceding round of scan-static refinement.
Another possibility is for the user code to take a sample of the
centrods before passing to the refiner, based for example on a
rough I/sigI calculation.