Modify strategy for determining the origin of the reference frame for a group of panels. Instead of first projecting the panel centres onto the parent frame, then taking the centroid of the projections, now first take the 3D centroids of the panel centres, then project this vector onto the parent frame.
This strategy is more robust when the children of a group are close to 90
degrees from the group frame, e.g. for the I23 barrel. When the group frame
and the children are approximately coplanar (e.g. the CSPAD) it makes very
little difference.

Either way, we rely on the parent frame for a group of panels being in some
way representative of the group. That is, it will fail if the vector pointing
to the centroid of the panels lies within the group frame. So, N.B. when
creating a hierarchical detector format the definition of parent groups
cannot not be arbitrary (maybe that should be documentation...)