New command dials.combine_experiments_and_reflections provides a simple way to combine separate pairs of experiments and reflections files into a single combined experiments file and a single combined reflections file. Experiment IDs are incremented according to the order of input.
Any (or none) of the input models may be chosen as a "reference" model for the
run. A reference model will replace all instances of that type of model in the
output experiment list. This makes it easy to produce a set of overlapping
experiments that share, e.g. the beam and detector models. This task is
frequently encountered when combining jobs from separate indexing runs for
multi-experiment refinement.