New command dials.slice_sweep
If provided experiments, takes subsets of these according to the new
image ranges set by (multiple) parameter scan_range definitions. That
is, it modifies the scan objects within those experiments.

If provided reflections, takes subsets of these so that only
reflections inside the image ranges given by the scan_range definitions
are kept. The first provided scan_range applies to reflections with
id=0, the second to id=1 and so on. Any reflections with ids outside
the range implied by the number of scan_range definitions will be
removed.

FIXME Needs a test!