Use logging in dials.refine and refinement modules.
Verbosity is now mostly handled by dials.util.log.config, except
engine.py still checks for verbosity > 2 to determine whether or
not to print curvatures (for LBFGS engines) and refiner.py
checks verbosity < 2 to decide whether to truncate the RMSDs by
experiments table. These might both be catered for instead by a
"super-debug" level.

Some notion of verbosity still seems useful to decide whether or
not to perform some expensive calculation that only results in
super-debug output, but an explicit verbosity variable should be
phased out in favour of checking the level of the current logger.