Make this test somewhat more useful and expose to run_tests.py.
run(verbose=True) to see summary output. Certain small subclasses of
of reflections have poor gradient calculations compared with finite
differences, for beam and crystal parameters. However for the majority
of reflections the agreement is good.

Perhaps this indicates a breakdown in the applicability of the
calculations, analogous to the 'too close to spindle' case for rotation
scans. This is not fully understood. In addition this test uses data
from a simulated narrow rotation scan in place of true stills data,
meaning that the 'observed' centroids are subject to systematic error.