This script attempts to answer how many single shot still images are needed for a 95% complete lysozyme dataset with 4 fold redunancy to 2 angstroms.  Random orientation matrices are generated, then given a detector and beam model and a simple mosaicity model, refelctions are predicted.  The reflections are added to a miller set.  The program exits when the above requirements are met and reports how many images are required.
A CSPAD CBF is used for the detecotor model to simulate the gaps
in the CSPAD.  This adds 50-100 images to the total.  A simple
CSPAD detector model is available, as is a MARCCD model.

Finally, the program attempts to determine the strength of the
anomalous signal using models provided by Tom Twilliger.

Run the program with no parameters:
libtbx.python completeness_sim.py

Future work:
-add I/sigma effects. This program assumes a redundancy of 4
is enough, after postrefinement, to completely correct intensities
to full reflection equivalents.  Therefore it reports an absolute
minimum number of images when the reality is almost certainly higher.