Fix long-standing bug in the Reeke index generator implementation that resulted in some indices that should be predicted failing to appear in the generated index list.
This is a legacy of the Mosflm implementation. There, the setting matrix
at the end of a frame is related to the setting matrix at the start of a
frame by a pure rotation. For that reason some values that do not depend
on the angle are not re-calculated at the end of frame.

In the DIALS implementation we relaxed the assumption of a pure rotation
between endpoints so that we can use the Reeke algorithm to generate
indices for a small segment of a scan-varying model. Unfortunately we
omitted to re-calculate values that do not depend on angle, but do
depend on the lattice metric.

This change ensures that quantities that depend on the setting are
always recalculated at the end of frame when calculating the Ewald
sphere limits. The resolution limits are still only calculated using the
beginning of frame setting matrix, however as a margin is added to the
limits, these are still always large enough to cover d_min for any
sensible use case for this algorithm.