Do not update minimum covering sphere
Updating the minimum covering sphere is a potentially expensive operation.
The run time depends on the random selection of points, as the scitbx
function itself notes. In practice run times of ~1 sec have been observed.
So calculate the sphere only once, and do not update when different sets
of reflections are selected. Fixes #53