More convenient interface for parameter fixing.
fix_list can now include partial parameter names as well as integer
indices to determine which parameters to fix. For example,

  crystal.orientation.fix_list=Phi1,Phi2

will fix the Phi1 and Phi2 parameters for all crystals.

  crystal.orientation.fix_list=Crystal1Phi1,Crystal2Phi1,Phi2

will fix Phi1 for the first two crystals and Phi2 for all crystals.

  detector.fix_list=Dist

will fix the distance for all panel groups of all detectors, while

  detector.fix_list=Detector1Group1Dist

will fix the distance for the first panel group of the first detector,
if this is a hierarchical detector model.

The parameter names can be determined by running dials.refine without
parameter fixing and looking in the dials.refine.debug.log.