dev.dials.diffraction_geometry_viewer
Development of a viewer for reciprocal space diffraction geometry intended
for exploration of improved stills prediction, taking into account profile
parameters.

Currently is simply reciprocal_lattice_viewer with an Ewald sphere.

In future, calculate and display predictions, angles of rotation to the ES
(round gonio axis if present or shortest rotation by DeltaPsi if not) and
reflections with size and shape using profile parameters.