Refactor dials.stills_process import step.
Problem: if indexing 1000s of still images in 1000s of files, we want to delay importing each image until right before that image is being processed by a worker thread to save on memory.  However, if working with 1000s of images in a single file (NeXus from Eiger, for example), need to pre-import the file to properly split images between threads.

Solution: if there is only one image provided, import it.  Then if it turns out to be a multi-image file, split it up between threads.  Otherwise, if the user provides > 1 file for still processing, delay import until mutiprocessing starts (i.e., make the iterable for parallel_map be the file names).

The user can override this behavior with dispatch.pre_import=True.  This would be done if the user has several multi-image files to process at once.  The imports will be done by the root process, one at a time, which can be slow.  Then, the images in the multi-image files are split up to all the threads.