Add parameters n_subset_method=*random n_refl and n_refl_panel_list
These parameters and their implementation are moved in from cspad.cbf_metrology.  If using n_subset, this adds an alternate method of choosing the subset instead of randomly, namely choosing the subset with the highest number of reflections.  This makes metrology refinement easier if choosing a subset from a large number of images.

n_refl_panel_list allow selection of which panels to search on when using n_subset_method=n_refl