Revert "dials.image_viewer: default gain=Auto, which implies to take it from the detector model".
This reverts commit f9e46fbbf8bdcc7ed11344bba8491dea55726783.

It turns out that the spot-finding gain set here does not replace the gain in the detector's panel object, but acts as a multiplier for it. Therefore this change was incorrect. We should expect gain=1.0 to give reasonable results in most cases, but it is useful to keep the parameter for fine-tuning the thresholding algorithm.