Minor scaling code changes include extra options and restructuring.
Able to select a subset of the reflection to determine the scale factors,
based on an I/sigma cutoff or dmin cutoff. Furthermore, the current
target_function code has a term to weakly restrain the scale factors to
one, this seems to improve the scale factors and Rpim for some datasets
(helps to avoid negative scales) but gives lhigher Rpim for other datasets
which require a larger variation in scale factors.