Scaling code updated to fix error in gradient calculation.
Gradient calculation is now correct but significantly slower. The max
iterations of the lbfgs minimiser has been set to 20 to avoid endless
minimisation. Modulation flag has been changed to a default value of
False for now. A finite difference gradient calculation is also in the
target function code, next this shall be moved to a separate 'test
derivates' script.