Add by_detector to dials.split_experiments, and update test
If True, instead of producing separate files for each experiment, experiments are grouped by unique detector model in the input set of experiments. For example, if there are five detector models in the input data, five sets of files will be produced, each containing experiments that reference a single detector model.