Add a CCP4-workshop rewrite of “Detail” tutorial.
This is using a newer dataset for consistency with the rest
of the workshop. Have taken the opportunity to try to rewrite
some of the gnarlier parts of the tutorial to be easier to
read and understand.

The tutorial-text-generator has also been updated to make
raw log files for this tutorial.