Change tutorial to match cache, and remove cache
Changed tutorial generation to match intention of 208050d5e - the cached copy
of the command was changed, but not the script that generated that.

Since the cached copies have drifted in several ways anyway, and are
regenerated when generating the rest of the tutorial data, have removed
the logs cache folder.