Change how the stills integrator handles weak data for stills
After integration, the stills integrator checks for reflections with non-positive variances, and fails the image if it finds any.  A reflection with a negative variance is a fail case and a reflection with zero error likely indicates a dead region where all the pixels have intensity zero.  This region is often behind a beamstop and should have been masked out.  However, for some detectors, a region with all pixels set to zero is fine, particularly with low dose counting detectors.  Therfore, instead of failing an image with reflections with zero error, check if those same reflections have zero signal.  If so, just remove them and move on.