dials.image_viewer refactoring
Nexus file loading took a very long time - ~10 minutes for 10k images. A
large proportion of this time was spent repeatedly interacting with wx and
GTK - in particular filling and fetching data from the wx.Choice/Combobox
- all image data and records were stored on the UI control, and data
management all done via the GUI.

This moves away from this 'design', to a plain(ish) image list on
the frame class. Because the data management was so tied to this control,
and because of the superclasses which are the basis for other viewers and
who knows what, a surrogate adaptor 'LegacyChooserAdapter' has been used 
in order to continue providing the essential parts of the interface to
these classes.

Both because the wx.Choice control doesn't offer a virtual interface,
and the general inappropriateness of a ComboBox for lists of 10k items,
an alternative, slider-based interface has been implemented.