Remove DIALS extensions, fixes #532.
This removes the existing extension mechanism based on abc
and metaclasses. Everything is now contained within the
/extensions/ directory and can be reached via eg.

  import dials.extensions
  plugin = dials.extensions.ProfileModel.load('plugin')

To add a new extension it must be imported and registered
in the relevant class in /extensions/__init__.py

One entry point is set up to allow loading experiments from
dxtbx, but entry points are not used in general at this time.

Co-authored-by: Markus Gerstel <markus.gerstel@diamond.ac.uk>