Modification to the behaviour of 2D/3D Gaussian Smoother.
Previously, the number of points to average over was set to
be the same for all dimensions, and was determined by the
minimum n_average of the dimensions, (using the calculation
from 1D smoother). i.e, for a 2D smoother parameterised by a
2 x 5 grid of points, the smoothing would be over two points in
x and y dimensions.
This has now been modified so that the smoothing can be different
in each dimension, determined by the number of parameters in each
dimension. i.e. for the example above, the smoothing would be
over 2 points in x and 3 points in y, (3 is the default number
of points to smooth over for a 1D smoother with 5 parameters).
This should probably have been the default behaviour when I
implemented the 2D and 3D smoothers, and I expect this to give
better behaviour for cases with a small number of parameters in
one or more of the dimensions.