Create string representation of object in __str__
Move away from the view that .show(out=) is the primary point of access
to a string representation of an object. This leads to us playing around
with StringIO() pseudo-files to extract the representation as a string,
which in most cases will be what we want. Because strings are versatile.
They can be written to files, written to logs *and* written to the screen.

So make __str__() create the string representation and if .show() is
really used for anything, then make that call __str__() in the first
instance.