Improve options for handling partials during scaling.
min_partiality= option is now used to select which reflections will
be used to minimise the scaling data, but not to set the
'excluded_for_scaling' flag.

partiality_cutoff= option added to allow cutting of the dataset below
a certain partiality (sets user_excluded_in_scaling flag - this flag
is used in export_mtz to cut the data and is also used in scaling to
set the bad_for_scaling flag, for which scale factors are not
calculated.

i.e. min_partiality=0.6, partiality_cutoff=0.2 - only reflections
with partiality > 0.6 will be used to determine the scaling model,
but inverse scale factors will be calculated for all reflections
with partiality > 0.2.