Change STDIN reading to require --phil option
OptionParser would previously attempt to read the STDIN by default,
which has caused problems in the past e.g. #466. This adds a default
--phil parameter to the OptionParser, and reads the file(s) specified
there. STDIN is still supported by the unix-standard practice of passing
'-' as a filename.

You can pass in --phil= multiple times to specify multiple separate
files.

This change was discussed and agreed a while ago, and @graeme-winter
agreed that it was still a useful change.

Closes #466.