Add a tqdm-based progress bar to dials.util
  >>> from dials.util import progress
  >>> for i in progress(range(10)):
  ...     ...

by default it will not be shown if the total time for the loop is
estimated to be shorter than 10 seconds (customizable with the min_time
parameter), and will not be shown at all for non-tty output. To avoid
problems with slow-starting loops, the bar will not be shows for the
first two seconds at all (unless min_time is set to <2).

See https://github.com/tqdm/tqdm for all other options and usage, for
which a few defaults have been changed:
  - dynamic_ncols is True by default
  - leave         is False by default
  - disable       is None by default (tty only)