Allow use of dials.geometry_viewer without images
Allow use of phil parameter require_images=False to skip the format
checking that requires images files etc for parsing the experimentlist.
Also print a help suggestion if an error is raised during parsing in
geometry viewer.

This means one no longer needs to use the environment variable
DIALS_EXPORT_DO_NOT_CHECK_FORMAT=FALSE, however the use of this
is still allowed for compatibility, but a warning message is printed.

Closes #484