Not-quite-finished changes for creating normal prob plot in dials.report
I originally wrote the intensity_explorer.py (probability plot) code to read in a scaled unmerged MTZ file. The entries in a scaled unmerged MTZ are labelled by symmetry-equivalent Miller index within the ASU, thus trivially permitting operations on reflections grouped by symmetry equivalence. If we want to do the same with reflections passed in as a reflection_table, we also need to pass in a corresponding ExperimentList, in order to obtain from it the crystal symmetry. If we are to incorporate normal probability plots, and associated other plots, into dials.report, we need to be able to call the probability plot code by passing it a reflection_table and ExperimentList, not by reading in an MTZ file.  A commit to fix this follows in ~7½ days.