dials.reindex: allow reindexing against a reference reflection file
This commit introduces an option to specify a reference reflections
and experiments file to dials.reindex, of the same space group as
the reflections and experiments to reindex. The space group is
tested for potential twinning operations. If these exist, the
different settings are tested, and the operation that gives the
best correlation to the reference is used to reindex the data.