refiner: Only copy parts of Experiment that are used
Cases where the Experiment can refer to large, complex objects means
that a deepcopy can be a very expensive operation - was seeing cases
where the copy took 6s because of e.g. large Eiger ImageSet objects.

However, the refiner still needs to actually change the experiment
objects that it is refining. This patch to the refiner makes copies
of the parts of the experiment that it refines, but leaves the other
parts of the Experiment as references to the original.

An earlier attempt at this fix was made in #259 however the usage of
this was quickly undone, I presume due to encapsulation problems. These
interface changes have been completely reverted as part of this commit.

Fixes #258.