Disentangle construction of parameterisation from refiner.
* Move PHIL parameters to parameterisation package
* Move config_restraints out of config_parameterisation

Move configuration of the parameterisation objects to its own module.

Remove the need for an isinstance call by subclassing Refiner.
Extra bits required for scan-varying refinement are now put in
a ScanVaryingRefiner. The refiner no longer needs a complicated
by-proxy method to work out whether it did scan-varying refinement,
as the RefinerFactory will have created the right kind of Refiner.