Persist array.info() property through symmetry_base
Although there are many places after these instances where the property
is explicitly copied, these two places don't. This caused the bug that
was fixed in 34fc7aa47acb4d714cb9845258c9eb7d4eb2769c, which effectively
ignored the fact that the copy semantic wasn't working properly.

I've written this in the style of the existing code but take no
responsibility for it - I don't know why e.g.

  self.intensities.set_sigma(sd)

wasn't used but am going to charitably assume that there was a reason or
strange behaviour of miller.array. I also assume it's far too late to
fix the inconsistent behaviour of array.customized_copy and that too much
depends on the current.