Port logging features from xfel: bread crumbs and per-process log files
Bread crumbs: simple sets of files that show the time stamps for step taken in processing every frame, including the outcome of the processing.  There will be a debug folder in the outputdirectory with one file per process.
Per-process log files: if logging_dir is specified, the logs will be split, one per rank.