Fix a memory leak in dials.stills_process. Affects composite image processing (hdf5 and xtc).
Leak came from the split_datablock code, code only used with hdf5 and FormatXTC processing.  Leak not applicable to processing many individual images like cbfs, or to processing xtc using cctbx.xfel.xtc_process.

Details: holding a reference to the split datablock means that when the data is read and cached by imageset.h, it never goes out of scope.  Instead, split the datablock in the worker processes, which means the split datablock goes out of scope and the cache is freed.

Thanks @RBolotovsky for working on this with me.