image_viewer: Fix display of integrated reflections
Commit ae4dbf started depending on the column imageset_id. It turns out that
integrated files don't have this column.

This commit tries to use the column, but falls back to a) The original
behaviour if only one ImageSet, or b) hiding unindexed reflections, because
there is no way to separate them via imageset.