New dials.refine default: reflections_per_degree = Auto.
This default was changed in 156dd8 from 100 to None, which implicitly
used all reflections for refinement. This gets slow for scan-varying
refinement of large multi-turn datasets, while previous tests showed
that refined models are essentially unaffected down to 100 ref/degree
(but below 50 ref/deg the quality of refinement results suffers).

This commit tunes the sample size for multi-turn datasets to reach
a minimum of 100 ref/degree. For <= single-turn datasets the default
is equivalent to the previous behaviour.