Resolving dials.export imageset_index semantics
Resolve the semantics question (#690) as follows:

If imageset_index is specified then the scan is treated as starting with
image number 1 and ending with image number len(scan). Therefore
dials.export_bitmaps will produce files in the format
{prefix}_{imageset_index}.{extension}, which corresponds to what the
user is most likely to expect in that instance, ie with

  dials.export_bitmaps data_0004.cbf imageset_index=1

we write the file image0001.png, since you asked for index number 1.

If imageset_index is not specified then the scan information is treated
as defined in the image headers, therefore dials.export_bitmaps will
produce files in the format {prefix}_{image_number}.{extension}, which
corresponds to what the user is most likely to expect in that instance,
ie with

  dials.export_bitmaps data_0004.cbf

we write the file image0004.png, since you asked to export that image
with no further qualification.

Resolves #690.