Use the median unit cell in scaling, or allow a best cell to be specified.
For multi-dataset analysis, in order to have consistent d-values for
applying a resolution cutoff or for giving the merging stats, a
consistent unit cell is needed across experiments (scaling was
previously using the unit cell of the first experiment). If a best
cell is specified, using the best_unit_cell= option (e.g. one could
give the result of dials.two_theta_refine), then this will be used,
else the median unit cell is determined and used.

Currently, this is only used when performing a resolution cut and
when calculating merging stats, the crystals in the experiments are
not updated, nor are the d-values in the reflection tables modified.