Kapton correction for single/multi-panel detectors
During our beamtime at SACLA in Feb, 2019, where the drop on tape (DOT) sample
delivery system was used, it was realized that a new tool is needed to correct
for the kapton tape shadow that affects integrated intensities. The existing
kapton tool in DIALS works only for single panel detectors. Here we present a
new implementation of the kapton correction making use of dxtbx objects. The
kapton tape itself is represented by 6 dxtbx detector objects (to form a
volume) and we have made use of various detector functions to get the path
length of the x-rays through the tape. Furthermore, no assumptions regarding
single/multi-panel detectors are necessary because the mathematics has been
coded up in a general way.

The code reproduces older kapton results on the Rayonix at LCLS using the same
set of parameters. More importantly, it now allows us to use it on multi-panel
detectors like the Octal-MPCCD.