Revert dropping cache
This needs some modification on the phil stuff to be able to import a
file directly, ie.

  dials.import some.cbf

Looks like we import the file, name it "<image files>" and when it comes
to the point of generating a diff phil we try to import the file named
"<image files>" again, which was intercepted by the caching layer.

Why we would attempt to import files when generating a diff-phil is
beyond me, so reverting for now.

This reverts commit 1d23379c338e48b97be408aa3cad074debbe9097.
This reverts commit 304a26cf1034a7505a54567a978ba5633793dc71.