Report standard error of correlation coefficients.
This is intended to address #215. Standard errors are printed
alongside the correlation coefficients. These can be large when
the number of reflections is small. To enact as a sanity check,
the correlation coefficient itself is not compared with
symop_threshold, but a lower bound given by (CC - SE) is
calculated first then this is compared with symop_threshold.