Restabilise test
This test appeared to get less stable after switching from randomly-
generated to pre-generated data. This is because the random seed was
no longer being set, but the mcd outlier rejection method is
stochastic. So, although the data was fixed the algorithm gave
variable results.

Set the random seed to stablise expected test behaviour and reenable
the skipped tests. Monitor behaviour for #836