Add new command line program dials.space_group
This program tests systematic absences based on the laue group,
to recommend a space group for the dataset, which is saved
to symmetrized.expt. Currently only tests MX space groups i.e.
tests screw axes but not glide planes and mirrors.

For each laue group, possible screw axes are scored individually,
based on a test of whether the expected absences and reflections
in the reflection condition are significantly different to zero,
based on <I>/<sigma> values. These scores are combined to give
a recommended space group. A html report is produced to inspect
the data.