dials.align_crystal: change results data structure
Originally this was an ordered dictionary with a 2-tuple of vectors as
key. Since vectors are not immutable they can't be used as keys in a
dictionary. Therefore the data structure is changed to a list of
3-tuples, which contain the 2 vectors and what was previously the
dictionary item value.

Change should only affect places where dials.command_line.align_crystal
is used programmatically, and applies to the .all_solutions property of
the object.

And yes, it's a Python 3 thing.