dials.align_crystal: change results data structure
I missed that .unique_solutions attempts to create an ordered set
with a 2-tuple of vectors, so the same issue and solution apply
here. The ordered set is changed to a list of values, the comparison
is done in O(n²), which is likely to be good enough.

Since the change does not affect the program output the same applies
as with the previous commit 7873537bc3406217ec212577f162214940859f83

Also fix a json writing bytes vs text python 3 issue.