Clean clutter
Remove large swathe of utterly irrelevant lines of data items from
PySlip init() - if you want to play battleships with this please pull
out an old version. Fixes unicode in source problem by removing the
offending unicode for Hong Kong. Happy side effect is that the code is
now a few hundred lines shorter. Begs the question of how this ever made
it in there. Latest versions of PySlip depend on wxPython 4 and Python 3
so currently cannot update.

Co-authored-by: Graeme Winter <graeme.winter@gmail.com>
Co-authored-by: Markus Gerstel <markus.gerstel@diamond.ac.uk>