dials.sweep_to_stills (#917)
* New program: dials.sweep_to_stills

Converts a sweep to a series of stills by using the goniometer and scan to create a stills Experiment for every scan point in the scan. Includes an oscillation_angle/2 offset to the A matrices to account for how scans are recorded in DIALS.

Also includes reflection splitting. Idea is that the indexed reflections from the sweep need to be split into slices so that each experiment gets all the spots that impinged on it, even if the centroid is on the next phi step.

Co-authored-by: Kevin Dalton <kevinmdalton@gmail.com>

* .rows() will do what you want here

* Code review

* Add simple test, add parameter max_scan_points, check for missing keys

* Reshuffle script to more standard form

- Split class __init__ and run() to more standard run (which loads the
  arguments from command line and writes the data back out) and
  sweep_to_stills (which transforms the data). This makes it easier to
  use programatically from code.
- Remove redundant set_dispatcher_name
- Remove shebang
- Use program name in help string directly
- use sys.exit so that exit code is set on bad input
- Use docstring instead of separate help_message

* Add a towncrier feature note

* Add parameter check_reference_geometry=True, expert_level=2, to dials.import

If False, skip the check reference detector step. Useful when the known headers are very wrong, such as the pixel size is off.