Use a subset of reflections for sigma_m calculation in integrate (#942)
In dials.integrate we used to use all reflections to calculate sigma_m using
a simplex minimiser, which takes a significant time for large datasets.
As this problem is hugely overdetermined, it should be possible to use a
subset of reflections and obtain a result that is just as accurate.
dials.refine already selects a well-sample set of the data, so use this subset
if available, but at least 10000 reflections.

Filter reflections at the top, use consistent set of reflections for
both sigma_b and sigma_m, as a courtesy report the number of reflections
used. Also separated out the still case for clarity, though means slight code
duplication occurs.