Minor change to dispersion extended spot finder.
Both the mean image display in dials.image_viewer and the
symmetry of the spot-finding algorithm are fixed. The former
by not overwriting the variable mean_ and the latter by
correcting an off-by-one error. Also adds a test to ensure
correct behaviour and removes the skip from tests added in
782e803e. Closes #946.

Co-authored-by: James Parkhurst <james.parkhurst@diamond.ac.uk>