Make reflection tables behave more like 'dict's (#908)
This changes parts of the reflection table interface.

The default iterator now iterates over columns instead of rows.
If iteration over rows is desired then existing code such as:

for r in reflections:
   r["bbox"] ...

for r1, r2 in zip(reflections1, reflections2):
   ...

needs to be amended to:

for r in reflections.rows():
   r["bbox"] ...

for r1, r2 in zip(reflections1.rows(), reflections2.rows()):
   ...

The methods .keys() and .items() behave as one would expect from a
dictionary-like object.

Attempts to access a non-existing key now results in a KeyError rather
than a RuntimeError.

With this the interface to flex.reflection_table is more consistent with
other similar containers such as a Python dictionary, and a pandas
DataFrame. For further details please see #908 and #862.