Replace libtbx.introspection.machine_memory_info with psutil
On MacOS, this function took 0.3 seconds to run per instantiation, which
quickly built up when running many times. It also appears not to work
properly cross-platform. psutil does this much faster (13 µs), and
supports all the platforms we do - so also removed the "sanity checks"
designed to catch bad implementations.

This speeds up indexing by ~17s for me for the first stage.

Also, a few driveby import floating changes. We don't need to import
math.ceil both in the top scope and in many scopes underneath.