For #994 - split experiments (#995)
* For #994 - split experiments

If we have > 1 scan for a given crystal model will need to duplicate this before moving on to scan varying refinement - split at the start of the processing to do scan static refinement with split crystals but shared detector, beam etc. Should accomodate > 1 detector but one crystal, one scan correctly as not splitting the experiment as was the previous behaviour. 

Remove the `split_experiments` option from dials.index and documentation.