Fix problem with conda dials environment file on maxos
Currently .conda-envs/dials_py27_osx-64.txt uses conda-forge::hdf5-external-filter-plugins.  When I run a simple bootstrap.py --builder=dials --use-conda on my mac, conda enters an endless dependency validation loop.  This is because conda-forge::hdf5-external-filter-plugins was compiled using the default clang compilers that use libcxx v. 9 or so, but the .conda-envs/dials_py27_osx-64.txt specifies cctbx::libcxx, which is version 4.0.1.  Thus the dependencies are conflicted and the installation fails.

This patch changes the env file for dials to use conda-forge::libcxx.

Having the hdf5 plugins on conda forge built using the latest conda supplied clang is a good thing so advancing to libcxx to the latest version makes sense.  However, this adds some risk.  I’ve ran libtbx.pytest --regression in the dials folder and that passed, so the basic functionality appears to be working.  We’ll let the azure/jenkins builders chew on this pull request and then maybe merge it tomorrow if it makes sense?  Likely we’ll need someone to test using a compressed h5 file.

(if this continues to work, we’ll likely migrate cctbx/phenix off of libcxx 4.0.1 onto the conda-forge version)