Putative fix for dials.generate_mask trusted_range bug (#978) (#980)
Ensure that generated masks do not include pixels that are overloaded
on a few images, but only pixels that are always outside the trusted range.

* Putative fix for #978

Loop is efficient for single-panel detectors and resolves the issue as published. Will be expensive for large image sets but that is the expected behaviour I would expect. Could do with refactoring for multi-panel detectors as the loop is inside out.

* Suggestion from @dagewa - if all of mask is True break

* Test that overloads don't affect the mask.

* Copy the phil scope on input

The input phil scope is later modified, which can cause subtle issues if the
option parser is called multiple times. Therefore create a copy of it on input
then we aren't modifying the original phil scope.

* Remove redundant line

* Convert procrunner tests to run function calls

* Add explanatory comment and newsfragment. Fixes #978

Co-authored-by: Graeme Winter <graeme.winter@gmail.com>
Co-authored-by: David Waterman <dagewa@users.noreply.github.com>