Include format_phil_scope in dials.export_bitmaps
This means that dynamic shadowing should be applied in the format class if
appropriate when given images as input rather than imported experiments.

The inclusion of the format scope meant that it was necessary to move the
format parameter a new output scope. Use this opportunity to move a few
other parameters into the output scope at the same time.