Allow controlling image_viewer via zmq
This adds a 'zmq_endpoint' PHIL parameter to image_viewer, that it binds a ZMQ
PULL socket to and listens for commands. At the moment, the only command is
to load a new image, and is a json messages of the form:

    {"command": "load_image", "image": "<filename>"}

This doesn't add a pyzmq dependency, so this will only work if explicitly
added to the distribution for now. It doesn't affect image_viewer at all
if pyzmq isn't present, but if you try to ask it to bind to an endpoint
without zeromq it'll exit with an error.

The primary usage for now has been for I24 serial online hitfinding, and
using dials.image_viewer to jump around the results.