Colour warning (and above) messages on the console (#1078)
Use the colorlog (https://github.com/borntyping/python-colorlog) package to colour messages written to the log console. Currently blue for DEBUG, yellow for WARNING, red for ERROR. This complements #1050, which marks up warning messages in the log _file_ with an automatic prefix. Now that warnings are highlighted in some way either on the console or the log file, remove extraneous parts of various warning messages throughout the codebase. Closes #1034