Fix shoebox memory usage check (#1121)
Sometimes dials.algorithms.integration.processor._average_shoebox_size can be called on a reflection table that has no shoeboxes. In fact, this is the usual case when dials.algorithms.integration.processor.Manager.compute_processors is called during a dials.integrate process because the predicted reflections have bboxes but don't yet have shoeboxes.

When the multiplicity is so high that the required shoebox memory exceeds the memory limit, dials.integrate should raise a nice RuntimeError with a meaningful message. This code path was broken and is fixed by this PR.

fixes #1111

Co-Authored-By: Markus Gerstel markus.gerstel@diamond.ac.uk